+++ src/main/java/webserver/Servlet.java
@@ -3,20 +3,6 @@ package webserver;
 import http.HttpRequest;
 import http.HttpResponse;
 
-public class Servlet implements HttpServlet {
-    @Override
-    public void doGet(HttpRequest request, HttpResponse response) {
-    }
-
-    @Override
-    public void doPost(HttpRequest request, HttpResponse response) {
-    }
-
-    @Override
-    public void doPut(HttpRequest request, HttpResponse response) {
-    }
-
-    @Override
-    public void doDelete(HttpRequest request, HttpResponse response) {
-    }
+public interface Servlet {
+    void service(HttpRequest request, HttpResponse response);
 }
